% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{boot.select.gen}
\alias{boot.select.gen}
\title{Perform a Step of FSS}
\usage{
boot.select.gen(
  resp.name,
  c.vars,
  myframe,
  type,
  r,
  f,
  model,
  fun.args,
  pred.args
)
}
\arguments{
\item{resp.name}{matrix of MPS data}

\item{c.vars}{matrix of MPS data}

\item{myframe}{data.frame holding explanatory and response data}

\item{type}{'reg' or 'class' for regression and classification. Not needed if `f` is specified.}

\item{r}{maximum cell count}

\item{f}{loss function}

\item{model}{name of modeling method; must be name of the function called in R}

\item{fun.args}{list of arguments that are passed to modeling function}

\item{pred.args}{list of arguments that are passed to model predict function}
}
\description{
Internal function that performs 1 step of forward stability selection. Returns a table of selection counts.
}
\examples{
library(MASS)
n<-100
p<-10
x<-mvrnorm(n,rep(0,p),diag(p))
signal<-rowSums(x[,1:3])
noise<-rnorm(n,0,1/4)
y<-signal+noise
mydata<-data.frame(x,y)
boot.select.gen(resp.name='y',c.vars=NULL,myframe=mydata,B=100,model='lm')
}
\keyword{matrix}
