% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{build.tree}
\alias{build.tree}
\title{Tree Graphic Function}
\usage{
build.tree(
  var.mat.na,
  cex = 0.9,
  alt = 1,
  shape1 = "trad",
  shape2 = "right",
  lwd = 1,
  col
)
}
\arguments{
\item{var.mat.na}{matrix of MPS data}

\item{cex}{text size}

\item{alt}{integer that vertically staggers terminal nodes (max of 3)}

\item{shape1}{either \code{'trad'} or \code{'rad'} for traditional trees and radial trees}

\item{shape2}{either \code{'right'} or \code{'tri'} for rectangular and triangular;
only matters for \code{shape1='trad'}}

\item{col}{node color}
}
\description{
Builds MPS graphic in graphics window
}
\examples{
mat_input<-c('Var1','Var1','Var1','Var1','Var2','Var2',
   'Var3','Var3','Var4','Var5','Var4','Var5')
mat1<-matrix(mat_input,nrow=4)
build.tree(mat1) #base example
build.tree(mat1,alt=2) #different alt
build.tree(mat1,shape1="rad") #radial tree
build.tree(mat1,shape2="tri") #triangular tree (only for shape1)
build.tree(mat1,col="red") #node color
}
\keyword{tree}
