% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{find_worstcases_chen86}
\alias{find_worstcases_chen86}
\title{Find set likely to contain the most probable category}
\usage{
find_worstcases_chen86(r, k, P)
}
\arguments{
\item{r}{maximum cell count}

\item{k}{number of cells}

\item{P}{desired inclusion probability}
}
\description{
Function that that finds the set likely to contain the most probable multinomial category
}
\examples{
find_worstcases_chen86(100, 5, 0.80)
}
