% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{full.select.gen}
\alias{full.select.gen}
\title{Perform MPS}
\usage{
full.select.gen(
  myframe,
  resp.name,
  depth,
  r,
  Pstar,
  f,
  type,
  model,
  fun.args,
  pred.args,
  condense = FALSE
)
}
\arguments{
\item{myframe}{data.frame holding explanatory and response data}

\item{resp.name}{matrix of MPS data}

\item{depth}{maximum depth of paths}

\item{r}{maximum cell count}

\item{Pstar}{inclusion probability bound}

\item{f}{loss function}

\item{type}{'reg' or 'class' for regression and classification. Not needed if \code{f} is specified.}

\item{model}{name of modeling method; must be name of the function called in R}

\item{fun.args}{list of arguments that are passed to modeling function}

\item{pred.args}{list of arguments that are passed to model predict function}

\item{condense}{logical for whether or not paths should be merged}
}
\description{
Function that performs Model Path Selection (MPS)
}
\examples{
set.seed(200)
library(MASS)
n<-1000
p<-10
x<-mvrnorm(n,rep(0,p),diag(p))
signal<-rowSums(x[,1:3])
noise<-rnorm(n,0,1/4)
y<-signal+noise
mydata<-data.frame(x,y)
mps1<-full.select.gen(myframe=mydata,resp.name='y',depth=3,
   B=100,model='lm',condense=FALSE)
mps1
build.tree(mps1) #graph

#merged paths
mps2<-full.select.gen(myframe=mydata,resp.name='y',depth=3,
   r=100,model='lm',condense=TRUE)
mps2
build.tree(mps2) #graph
}
\keyword{matrix}
