% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{naAdder}
\alias{naAdder}
\title{Add NAs to MPS Graph Matrix}
\usage{
naAdder(myMat)
}
\arguments{
\item{myMat}{matrix of MPS data}
}
\description{
Internal function that adds \code{NA}s to matrix that holds MPS graph data
}
\examples{
mat_input<-c('Var1','Var1','Var1','Var1','Var2','Var2',
   'Var3','Var3','Var4','Var5','Var4','Var5')
mat1<-matrix(mat_input,nrow=4)
naAdder(mat1)
}
\keyword{matrix}
