% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{naRemover}
\alias{naRemover}
\title{Remove NAs to MPS Graph Matrix}
\usage{
naRemover(myMat)
}
\arguments{
\item{myMat}{matrix of MPS data}
}
\description{
Internal function that removes \code{NA}s to matrix that holds MPS graph data
}
\examples{
mat_input<-c('Var1',NA,NA,NA,'Var2',NA,
   'Var3',NA,'Var4','Var5','Var4','Var5')
mat1<-matrix(mat_input,nrow=4)
naRemover(mat1)
}
\keyword{matrix}
