% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{read.tree.label}
\alias{read.tree.label}
\title{Tree Processing Function}
\usage{
read.tree.label(var.mat.na)
}
\arguments{
\item{var.mat.na}{matrix of MPS data}
}
\description{
This is an internal function used for building MPS graph.
}
\keyword{matrix}
