% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tree.R
\name{read_tree_label}
\alias{read_tree_label}
\title{Tree Processing Function}
\usage{
read_tree_label(var.mat.na)
}
\arguments{
\item{var.mat.na}{matrix of MPS data}
}
\description{
This is an internal function used for building MPS graph.
}
\keyword{matrix}
