% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{wc_updater_finder}
\alias{wc_updater_finder}
\title{Update table}
\usage{
wc_updater_finder(r, k, P)
}
\arguments{
\item{r}{maximum cell count}

\item{k}{number of cells}

\item{P}{desired inclusion probability}
}
\description{
Function that that updates a table containing info generated by `find_worstcases_chen86()`
}
\examples{
find_worstcases_chen86(100. 5, 0.80)
}
