% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_updater_finder.R
\name{wc_updater_finder}
\alias{wc_updater_finder}
\title{Update table}
\usage{
wc_updater_finder(r, k, P, save.table = F, table.path = NULL, mc.cores = NULL)
}
\arguments{
\item{r}{maximum cell count}

\item{k}{number of cells}

\item{P}{desired inclusion probability}

\item{save.table}{whether you want to save a table for faster computation}

\item{table.path}{where table gets saved; only matters if `save.table=T`}
}
\description{
Function that that updates a table containing info generated by `find_worstcases_chen86()`
}
